abagabe fertig

+++ src/main/java/task_05_Adapter/Adapter.java
@@ -6,7 +6,7 @@ public class Adapter extends Box implements IChargingStation{
     }
     @Override
     public void chargStation(int amount) {
-        int pole = (int) amount/3;
+        int pole = amount /3;
         this.charge(pole,pole,pole);
     }
 }
